Add tests for login component
