+++ package.json
@@ -89,6 +89,7 @@
         "react-redux": "^5.0.6",
         "react-router-dom": "^5.1.2",
         "redux": "^4.0.0",
+        "redux-mock-store": "^1.5.4",
         "seamless-immutable": "^7.1.4"
     },
     "devDependencies": {
